requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
dj-database-url==0.4.2
Django==1.11.4
django-filter==1.0.4
django-test-without-migrations==0.6
djangorestframework==3.6.3
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
pkg-resources==0.0.0
python-decouple==3.0
pytz==2017.2
requests==2.18.3
uritemplate==3.0.0
urllib3==1.22
